requirements.txt
Django==4.0.0
pymongo==4.6.3
django-rest-framework
psycopg2-binary
python-dotenv==1.0.1
gunicorn
uvicorn[standard]